third_party/WebKit/LayoutTests/imagecapture/MediaStreamTrack-applyConstraints.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/mock-imagecapture.js"></script>
<body>
<canvas id='canvas' width=10 height=10/>
</body>
<script>

const meteringModeNames = ["none", "manual", "single-shot", "continuous"];

// This test verifies that we can all MediaStreamTrack.applyConstraints(), with
// a mock Mojo service implementation.

async_test(function(t) {
  var canvas = document.getElementById('canvas');
  var context = canvas.getContext("2d");
  context.fillStyle = "red";
  context.fillRect(0, 0, 10, 10);

  const constraints = { advanced : [{ whiteBalanceMode : "manual",
                                      exposureMode     : "continuous",
                                      focusMode        : "single-shot",

                                      pointsOfInterest : [{x : 0.1, y : 0.2},
                                                          {x : 0.3, y : 0.4}],

                                      exposureCompensation : 133.77,
                                      colorTemperature     : 6000,
                                      iso                  : 120.0,

                                      brightness           : 3,
                                      contrast             : 4,
                                      saturation           : 5,
                                      sharpness            : 6,

                                      zoom                 : 3.141592,

                                      torch                : true
                                    }]};

  var theMock = null;
  var videoTrack = null;
  mockImageCaptureReady
    .then(mock => {
      theMock = mock;
      var stream = canvas.captureStream();
      videoTrack = stream.getVideoTracks()[0];

      // |videoTrack|'s capabilities gathering, just like the actual capture, is
      // a process kicked off right after creation, we introduce a small delay
      // to allow for those to be collected, since they are needed to understand
      // which constraints are supported in applyConstraints().
      // TODO(mcasas): this shouldn't be needed, https://crbug.com/711524.
      return new Promise(resolve => setTimeout(resolve, 100));
    },
    error => {
      assert_unreached("Error creating MockImageCapture: " + error);
    })
    .then(function() {
      return videoTrack.applyConstraints(constraints);
    })
    .then(function() {
      assert_equals(constraints.advanced[0].whiteBalanceMode,
                    meteringModeNames[theMock.options().white_balance_mode],
                    'whiteBalanceMode');
      assert_equals(constraints.advanced[0].exposureMode,
                    meteringModeNames[theMock.options().exposure_mode],
                    'exposureMode');
      assert_equals(constraints.advanced[0].focusMode,
                    meteringModeNames[theMock.options().focus_mode],
                    'focusMode');

      assert_equals(constraints.advanced[0].pointsOfInterest.length,
                    theMock.options().points_of_interest.length,
                    'pointsOfInterest');
      for (i = 0; i < constraints.advanced[0].pointsOfInterest.length; ++i) {
        assert_approx_equals(constraints.advanced[0].pointsOfInterest[i].x,
                             theMock.options().points_of_interest[i].x, 0.01,
                             'pointsOfInterest x');
        assert_approx_equals(constraints.advanced[0].pointsOfInterest[i].y,
                             theMock.options().points_of_interest[i].y, 0.01,
                             'pointsOfInterest y');
      }

      assert_equals(constraints.advanced[0].exposureCompensation,
                    theMock.options().exposure_compensation,
                    'exposureCompensation');
      assert_equals(constraints.advanced[0].colorTemperature,
                    theMock.options().color_temperature, 'colorTemperature');
      assert_equals(constraints.advanced[0].iso, theMock.options().iso, 'iso');

      assert_equals(constraints.advanced[0].brightness,
                    theMock.options().brightness, 'brightness');
      assert_equals(constraints.advanced[0].contrast,
                    theMock.options().contrast, 'constrast');
      assert_equals(constraints.advanced[0].saturation,
                    theMock.options().saturation, 'saturation');
      assert_equals(constraints.advanced[0].sharpness,
                    theMock.options().sharpness, 'sharpness');

      assert_equals(constraints.advanced[0].torch, theMock.options().torch,
                    'torch');

      t.done();
    })
    .catch(error => {
      assert_unreached("applyConstraints(): " + error.message);
    })

}, 'exercises MediaStreamTrack.applyConstraints(constraints)');

</script>
